[Engine] Use int64_t for Lua round-trip handles and enum values; add lua_narrow to clamp and report out-of-range Lua integers

+++ src/engine/components/input_component.h
@@ -4,7 +4,7 @@
 #include "engine/core/systems/input.h"
 
 namespace hob {
-    using BindingId = int32_t;
+    using BindingId = int64_t;
     using AxisBindingFunc = std::function<void(float)>;
     using ActionBindingFunc = std::function<void()>;
